requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.3.1
    # via
    #   chex
    #   optax
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via shiny
appnope==0.1.4
    # via ipykernel
asgiref==3.8.1
    # via shiny
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   mbi
    #   referencing
beautifulsoup4==4.13.4
    # via nbconvert
black==25.1.0
    # via dp-wizard-templates
bleach[css]==6.2.0
    # via nbconvert
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
chex==0.1.90
    # via
    #   mbi
    #   optax
click==8.1.8
    # via
    #   black
    #   shiny
    #   uvicorn
comm==0.2.2
    # via ipykernel
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
deprecated==1.2.18
    # via opendp
dp-wizard-templates==0.7.0
    # via -r requirements.in
exceptiongroup==1.3.0
    # via
    #   anyio
    #   ipython
executing==2.2.0
    # via stack-data
faicons==0.2.2
    # via -r requirements.in
fastjsonschema==2.21.1
    # via nbformat
fonttools==4.61.0
    # via
    #   -r requirements.in
    #   matplotlib
h11==0.16.0
    # via uvicorn
htmltools==0.6.0
    # via
    #   faicons
    #   shiny
idna==3.10
    # via
    #   anyio
    #   requests
ipykernel==6.29.5
    # via dp-wizard-templates
ipython==8.35.0
    # via ipykernel
jax==0.6.2
    # via
    #   chex
    #   mbi
    #   optax
jaxlib==0.6.2
    # via
    #   chex
    #   jax
    #   mbi
    #   optax
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via nbconvert
joblib==1.4.2
    # via scikit-learn
jsonschema==4.23.0
    # via nbformat
jsonschema-specifications==2024.10.1
    # via jsonschema
jupyter-client==8.6.3
    # via
    #   dp-wizard-templates
    #   ipykernel
    #   nbclient
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.3.0
    # via nbconvert
jupytext==1.17.0
    # via dp-wizard-templates
kiwisolver==1.4.8
    # via matplotlib
linkify-it-py==2.0.3
    # via shiny
markdown-it-py==3.0.0
    # via
    #   jupytext
    #   mdit-py-plugins
    #   shiny
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.10.1
    # via -r requirements.in
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mbi==1.0.0
    # via opendp
mdit-py-plugins==0.4.2
    # via
    #   jupytext
    #   shiny
mdurl==0.1.2
    # via markdown-it-py
mistune==3.1.3
    # via nbconvert
ml-dtypes==0.5.3
    # via
    #   jax
    #   jaxlib
mypy-extensions==1.1.0
    # via black
narwhals==1.35.0
    # via shiny
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.6
    # via dp-wizard-templates
nbformat==5.10.4
    # via
    #   jupytext
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.4.2
    # via mbi
numpy==2.2.4
    # via
    #   chex
    #   contourpy
    #   jax
    #   jaxlib
    #   matplotlib
    #   mbi
    #   ml-dtypes
    #   opendp
    #   optax
    #   pandas
    #   randomgen
    #   scikit-learn
    #   scipy
opendp[mbi]==0.14.1
    # via -r requirements.in
opt-einsum==3.4.0
    # via jax
optax==0.2.5
    # via mbi
orjson==3.10.16
    # via shiny
packaging==24.2
    # via
    #   black
    #   htmltools
    #   ipykernel
    #   jupytext
    #   matplotlib
    #   nbconvert
    #   shiny
pandas==2.3.2
    # via mbi
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via
    #   -r requirements.in
    #   matplotlib
platformdirs==4.3.7
    # via
    #   black
    #   jupyter-core
polars==1.32.0
    # via opendp
prompt-toolkit==3.0.51
    # via
    #   ipython
    #   questionary
    #   shiny
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==19.0.1
    # via opendp
pygments==2.19.1
    # via
    #   ipython
    #   nbconvert
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-multipart==0.0.20
    # via shiny
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via jupytext
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
questionary==2.1.0
    # via shiny
randomgen==2.1.1
    # via opendp
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via -r requirements.in
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.6.1
    # via opendp
scipy==1.15.2
    # via
    #   jax
    #   jaxlib
    #   mbi
    #   scikit-learn
shiny==1.4.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
starlette==0.49.1
    # via
    #   -r requirements.in
    #   shiny
threadpoolctl==3.6.0
    # via scikit-learn
tinycss2==1.4.0
    # via bleach
tomli==2.2.1
    # via
    #   black
    #   jupytext
toolz==1.0.0
    # via chex
tornado==6.5.1
    # via
    #   -r requirements.in
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.13.2
    # via
    #   anyio
    #   asgiref
    #   beautifulsoup4
    #   black
    #   chex
    #   exceptiongroup
    #   htmltools
    #   ipython
    #   mistune
    #   referencing
    #   shiny
    #   starlette
    #   uvicorn
tzdata==2025.2
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.6.0
    # via
    #   -r requirements.in
    #   requests
uvicorn==0.34.1
    # via shiny
watchfiles==1.0.5
    # via shiny
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websockets==15.0.1
    # via shiny
wrapt==1.17.2
    # via deprecated
